spack.mirror: split into submodules (#47936)
